refactor(project): Decimal support added

+++ library/src/main/java/com/github/ushiosan23/android_utilities/android/time/TimeUnits.kt
@@ -9,7 +9,7 @@ object TimeUnits {
 	 * @param v Total time unit
 	 */
 	@JvmStatic
-	fun getUnit(v: Number = 1): Long = 1 * v.toLong()
+	fun getUnit(v: Number = 1): Long = (1 * v.toDouble()).toLong()
 
 	/**
 	 * Get milliseconds unit
@@ -17,7 +17,7 @@ object TimeUnits {
 	 * @param v Total seconds
 	 */
 	@JvmStatic
-	fun getSeconds(v: Number = 1): Long = getUnit(1000) * v.toLong()
+	fun getSeconds(v: Number = 1): Long = (getUnit(1000) * v.toDouble()).toLong()
 
 	/**
 	 * Get milliseconds unit
@@ -25,7 +25,7 @@ object TimeUnits {
 	 * @param v Total minutes
 	 */
 	@JvmStatic
-	fun getMinutes(v: Number = 1): Long = getSeconds(60) * v.toLong()
+	fun getMinutes(v: Number = 1): Long = (getSeconds(60) * v.toDouble()).toLong()
 
 	/**
 	 * Get milliseconds unit
@@ -33,7 +33,7 @@ object TimeUnits {
 	 * @param v Total hours
 	 */
 	@JvmStatic
-	fun getHours(v: Number = 1): Long = getMinutes(60) * v.toLong()
+	fun getHours(v: Number = 1): Long = (getMinutes(60) * v.toDouble()).toLong()
 
 	/**
 	 * Get milliseconds unit
@@ -41,6 +41,6 @@ object TimeUnits {
 	 * @param v Total days
 	 */
 	@JvmStatic
-	fun getDays(v: Number = 1): Long = getHours(24) * v.toLong()
+	fun getDays(v: Number = 1): Long = (getHours(24) * v.toDouble()).toLong()
 
 }
